//Mahesh

+++ src/main/java/com/ats/webapi/model/ServDetailForPdf.java
@@ -32,10 +32,10 @@ public class ServDetailForPdf {
 	private int sprId;
 
 	@Column(name = "spr_qty")
-	private int sprQty;
+	private float sprQty;
 
 	@Column(name = "spr_rate")
-	private int sprRate;
+	private float sprRate;
 	
 	@Column(name = "spr_taxable_amt")
 	private float sprTaxableAmt;
@@ -115,19 +115,19 @@ public class ServDetailForPdf {
 		this.sprId = sprId;
 	}
 
-	public int getSprQty() {
+	public float getSprQty() {
 		return sprQty;
 	}
 
-	public void setSprQty(int sprQty) {
+	public void setSprQty(float sprQty) {
 		this.sprQty = sprQty;
 	}
 
-	public int getSprRate() {
+	public float getSprRate() {
 		return sprRate;
 	}
 
-	public void setSprRate(int sprRate) {
+	public void setSprRate(float sprRate) {
 		this.sprRate = sprRate;
 	}
